adding confirmation prompt

+++ nuke.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
-rm -rf ~/.config/
-rm -rf ~/.cache/
-rm -rf ~/.ssh/
+read -p "Are you sure? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  rm -rf ~/.config/
+  rm -rf ~/.cache/
+  rm -rf ~/.ssh/
+fi
+
